+++ dependency/coordinates/include/coordinates/coordinates.hpp
@@ -1,0 +1,12 @@
+#pragma once
+#include "aer.hpp"
+#include "ecef.hpp"
+#include "ecef_aer.hpp"
+#include "ecef_enu.hpp"
+#include "ecef_llh.hpp"
+#include "ecef_ned.hpp"
+#include "eci.hpp"
+#include "ellipsoid.hpp"
+#include "enu.hpp"
+#include "llh.hpp"
+#include "ned.hpp"